add spectification and add product function

+++ src/main/java/com/task_haibazo/entity/Product.java
@@ -58,7 +58,7 @@ public class Product {
 	@Column(name = "discount")
 	@Min(0)
     @Max(100)
-	private int discount;
+	private Double discount;
 	
 	@Lob
 	@Column(name = "image", nullable = false)
